Add files via upload
To share the full project folder, I had to break it into pieces. For the website to work properly again, open the zip folders and extract all 4 of the "GroupProjectX" folders into the main folder with the index.html and logo pngs. If that doesn't work, I'll try something else